add home page and enhance ui

+++ src/main/java/com/turtlesltd/sothikbhara/fare/cng/CngFareRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CngFareRepository extends JpaRepository<CngFare, Integer> {
+public interface CngFareRepository extends JpaRepository<CngFare, Long> {
     List<CngFare> findAllByUserId(Long user_id);
     void deleteByIdAndUser_Id(Long id, Long user_id);
 }
